Backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-multipart==0.0.6
openai==0.27.8
firebase-admin==6.2.0
python-dotenv==1.0.0
werkzeug==2.3.7
